Coverage Plot for MDEI Object

Coverage Plot for MDEI Object

coverPlot(
  object,
  xvar = "treat",
  sigval = 0,
  target = "tau",
  colors = c(gray(0.7), gray(0)),
  cex.point = 0.5,
  xlabel = "",
  ylabel = "",
  ...
)

Arguments

object

An object of class MDEI.

xvar

The variable to plot along the x-axis. May be `treat` for the treatment variable, the name of a column in the covariate matrix in the obj, or a numeric vector, the length of the data. Default is "treat".

sigval

The value to see if it is covered by the conformal interval. Either a single value or a vector. Default is 0.

target

Either tau` or `theta`. The first, `tau`, is the marginal effect at each point, while the second `theta`, is the portion of the conditional mean of the outcome that is a function of the treatment and the outcome.

colors

A vector with two elements. The first is the color of the confidence interval at points where the conformal interval does not contain sigval, the second the color where it does. Default is c(gray(.7), gray(0)).

cex.point

The size of the points in the figure. Default is 0.5.

xlabel

Label for x-axis of figure. Default is "".

ylabel

Label for y-axis of figure. Default is "".

...

Additional arguments to be passed to plot.

Value

No return value.